refactor: Remove GithubRepositoryClient class
