+++ src/analyzer/circleci_analyzer.ts
@@ -1,7 +1,7 @@
 import { sumBy, min, max, sortBy, first, last } from "lodash"
 import { Status, diffSec, Analyzer, secRound, TestReport, WorkflowParams, convertToReportTestSuites } from "./analyzer"
 import { WorkflowRun, SingleBuildResponse, CircleciStatus, TestResponse } from "../client/circleci_client"
-import { RepositoryTagMap } from "../client/github_repository_client"
+import { RepositoryTagMap } from "../client/github_client"
 import { TestSuite, TestCase } from "junit2json"
 
 type WorkflowReport = {
